fix cyclic dependecy and added integration test you can simply run to test aagainst your own device
