+++ .gitignore
@@ -117,3 +117,6 @@ PyViCare/__init__.py
 
 .DS_Store
 .vscode/settings.json
+
+token.save
+env
