+++ PyViCare/PyViCare.py
@@ -1,50 +1,9 @@
-from PyViCare.PyViCareDevice import Device
+from PyViCare.PyViCareDeviceConfig import PyViCareDeviceConfig
 from PyViCare.PyViCareOAuthManager import ViCareOAuthManager
-from PyViCare.PyViCareGazBoiler import GazBoiler
-from PyViCare.PyViCareFuelCell import FuelCell
-from PyViCare.PyViCareHeatPump import HeatPump
-from PyViCare.PyViCareOilBoiler import OilBoiler
-from PyViCare.PyViCarePelletsBoiler import PelletsBoiler
-from PyViCare.PyViCareService import ViCareDeviceAccessor
+from PyViCare.PyViCareService import ViCareDeviceAccessor, ViCareService
 from PyViCare.PyViCareCachedService import ViCareCachedService
-from PyViCare.PyViCareService import ViCareService
 
-
-class PyViCareDeviceConfig:
-    def __init__(self, service, device_model, status):
-        self.service = service
-        self.device_model = device_model
-        self.status = status
-
-    def asGeneric(self):
-        return Device(self.service)
-
-    def asGazBoiler(self):
-        return GazBoiler(self.service)
-
-    def asFuelCell(self):
-        return FuelCell(self.service)
-
-    def asHeatPump(self):
-        return HeatPump(self.service)
-
-    def asOilBoiler(self):
-        return OilBoiler(self.service)
-
-    def asPelletsBoilder(self):
-        return PelletsBoiler(self.service)
-
-    def getConfig(self):
-        return self.service.accessor
-
-    def getModel(self):
-        return self.device_model
-
-    def isOnline(self):
-        return self.status == "Online"
-
-
-class ViCare:
+class PyViCare:
     def __init__(self):
         self.cacheDuration = 60
