+++ PyViCare/PyViCareDeviceConfig.py
@@ -1,0 +1,42 @@
+from PyViCare.PyViCareDevice import Device
+from PyViCare.PyViCareGazBoiler import GazBoiler
+from PyViCare.PyViCareFuelCell import FuelCell
+from PyViCare.PyViCareHeatPump import HeatPump
+from PyViCare.PyViCareOilBoiler import OilBoiler
+from PyViCare.PyViCarePelletsBoiler import PelletsBoiler
+
+
+class PyViCareDeviceConfig:
+    def __init__(self, service, device_model, status):
+        self.service = service
+        self.device_model = device_model
+        self.status = status
+
+    def asGeneric(self):
+        return Device(self.service)
+
+    def asGazBoiler(self):
+        return GazBoiler(self.service)
+
+    def asFuelCell(self):
+        return FuelCell(self.service)
+
+    def asHeatPump(self):
+        return HeatPump(self.service)
+
+    def asOilBoiler(self):
+        return OilBoiler(self.service)
+
+    def asPelletsBoilder(self):
+        return PelletsBoiler(self.service)
+
+    def getConfig(self):
+        return self.service.accessor
+
+    def getModel(self):
+        return self.device_model
+
+    def isOnline(self):
+        return self.status == "Online"
+
+
