--- PyViCare/__init__.py
@@ -1,2 +1,0 @@
-__all__ = ['PyViCareService', 'PyViCareDevice', 'PyViCareGazBoiler',
-           'PyViCareHeatPump', 'PyViCareCachedService', 'PyViCareOilBoiler', 'PyViCareService', 'PyViCare']
